requirements.txt
amqp==5.0.6
asgiref==3.3.1
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-3-jet==1.0.8
django-cors-headers==3.7.0
django-countries==7.0
django-filter==2.4.0
django-modeltranslation==0.16.2
django-rest-resetpassword==0.1.2
djangorestframework==3.12.4
idna==2.10
kombu==5.0.2
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.1.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
